- change components structure - add dedicated user info component into header - prepare header for profile link redirect
