+++ apps/todoapp/src/app/todos/components/header/header.component.html
@@ -1,11 +1,4 @@
 <header class="header">
-  <h1>todos</h1>
-  <input
-    class="new-todo"
-    placeholder="What needs to be done?"
-    auto
-    [value]="text"
-    (keyup)="changeText($event)"
-    (keyup.enter)="addTodo()"
-  />
+  <todolist-user-info></todolist-user-info>
+  <!--Add Login/Logout button based on checkAuthStatus()-->
 </header>
