+++ apps/todoapp/src/app/todos/components/header/header.component.scss
@@ -1,0 +1,8 @@
+.header {
+  background-color: #b83f45;
+  color: #fff;
+  padding: 15px 20px;
+  display: flex;
+  align-items: center;
+  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.1);
+}
