+++ apps/todoapp/src/app/todos/components/header/header.component.ts
@@ -1,22 +1,8 @@
 import { Component } from '@angular/core';
-import { TodosService } from '../../services/todos.service';
 
 @Component({
   selector: 'app-todos-header',
   templateUrl: './header.component.html',
+  styleUrls: ['./header.component.scss'],
 })
-export class HeaderComponent {
-  text = '';
-
-  constructor(private todoService: TodosService) {}
-
-  changeText(event: Event) {
-    const target = event.target as HTMLInputElement;
-    this.text = target.value;
-  }
-
-  addTodo(): void {
-    this.todoService.addTodo(this.text);
-    this.text = '';
-  }
-}
+export class HeaderComponent {}
