+++ apps/todoapp/src/app/todos/components/todos/todos.component.html
@@ -1,6 +1,5 @@
-<div class="user">{{ user.username }}</div>
+<app-todos-header></app-todos-header>
 <div class="todoapp">
-  <app-todos-header></app-todos-header>
   <app-todos-main></app-todos-main>
-  <app-todos-footer></app-todos-footer>
 </div>
+<app-todos-footer></app-todos-footer>
