--- apps/todoapp/src/app/todos/components/todos/todos.component.scss
@@ -1,4 +1,0 @@
-.user {
-  display: flex;
-  justify-content: flex-end;
-}
