+++ apps/todoapp/src/app/todos/components/user-info/user-info.component.html
@@ -1,0 +1,4 @@
+<div class="user">
+  <p>Welcome</p>
+  <a [routerLink]="['/profile/' + user.uuid]"> {{ user.username }}</a>
+</div>
