+++ apps/todoapp/src/app/todos/components/user-info/user-info.component.scss
@@ -1,0 +1,8 @@
+.user {
+  display: flex;
+  align-items: center;
+}
+
+.user p {
+  margin-right: 8px;
+}
