+++ apps/todoapp/src/app/todos/modules/main.module.ts
@@ -1,0 +1,31 @@
+import { NgModule } from '@angular/core';
+import { TodosComponent } from '../components/todos/todos.component';
+import { Routes, RouterModule } from '@angular/router';
+import { HeaderComponent } from '../components/header/header.component';
+import { TodosService } from '../services/todos.service';
+import { MainComponent } from '../components/main/main.component';
+import { CommonModule } from '@angular/common';
+import { TodoComponent } from '../components/todo/todo.component';
+import { FooterComponent } from '../components/footer/footer.component';
+import { UserInfoComponent } from '../components/user-info/user-info.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: TodosComponent,
+  },
+];
+
+@NgModule({
+  declarations: [
+    TodosComponent,
+    HeaderComponent,
+    UserInfoComponent,
+    MainComponent,
+    TodoComponent,
+    FooterComponent,
+  ],
+  imports: [CommonModule, RouterModule.forChild(routes)],
+  providers: [TodosService],
+})
+export class MainModule {}
